Began parser renderpass construction

+++ project/GFlow_Editor/src/metaresources/Renderpass.cpp
@@ -28,6 +28,6 @@ void RenderpassResource::addConnection(const size_t left_uid, const size_t left_
     }
     Connection** connection = (*connections).emplace_back(true);
     (*connection)->getFirst()->setValues(left_uid, left_pin);
-    (*connection)->getSecond()->setValues(right_uid, right_pin);
+    (*connection)->getSecond()->setValues(right_uid, static_cast<int>(right_pin));
 }
